# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API-сервер для управления заявками и факторами риска в системе FRAX.",
        "title": "API для системы FRAX",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8090",
    "basePath": "/api",
    "paths": {
...
        "/frax": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает отфильтрованный список всех сформированных заявок (кроме черновиков и удаленных).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "frax"
                ],
                "summary": "Получить список заявок",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Фильтр по статусу заявки",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по дате 'от' (формат YYYY-MM-DD)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по дате 'до' (формат YYYY-MM-DD)",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ds.FraxDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/frax/cart": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает ID черновика текущего пользователя и количество факторов в нем.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "frax"
                ],
                "summary": "Получить информацию для иконки корзины",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.CartBadgeDTO"
                        }
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/frax/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает полную информацию о заявке, включая привязанные факторы.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "frax"
                ],
                "summary": "Получить одну заявку по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.FraxDTO"
                        }
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Заявка не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Позволяет пользователю обновить поля своей заявки (возраст, пол, вес, рост).",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "frax"
                ],
                "summary": "Обновить данные заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "updateData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.FraxUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Логически удаляет заявку, переводя ее в статус \"удалена\".",
                "tags": [
                    "frax"
                ],
                "summary": "Удалить заявку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/frax/{id}/factors/{factor_id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Изменяет дополнительное описание для конкретного фактора в рамках одной заявки.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "frax"
                ],
                "summary": "Обновить описание фактора в заявке",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID фактора",
                        "name": "factor_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новое описание",
                        "name": "updateData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.FactorToFraxUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет связь между заявкой и фактором.",
                "tags": [
                    "frax"
                ],
                "summary": "Удалить фактор из заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID фактора",
                        "name": "factor_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/frax/{id}/form": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Переводит заявку из статуса \"черновик\" в \"сформирована\".",
                "tags": [
                    "frax"
                ],
                "summary": "Сформировать заявку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки (черновика)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Не все поля заполнены",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/frax/{id}/resolve": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Модератор завершает (с расчетом) или отклоняет заявку.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "frax"
                ],
                "summary": "Завершить или отклонить заявку (только модератор)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Действие: 'complete' или 'reject'",
                        "name": "action",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.FraxResolveRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ds.CartBadgeDTO": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "frax_id": {
                    "type": "integer"
                }
            }
        },
...
        "ds.FactorInFraxDTO": {
            "type": "object",
            "properties": {
                "argument": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "factor_id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "ds.FactorToFraxUpdateRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                }
            }
        },
...
        "ds.FraxDTO": {
            "type": "object",
            "properties": {
                "PHF": {
                    "type": "number"
                },
                "POF": {
                    "type": "number"
                },
                "age": {
                    "type": "integer"
                },
                "complition_date": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "creator_login": {
                    "type": "integer"
                },
                "factors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.FactorInFraxDTO"
                    }
                },
                "forming_date": {
                    "type": "string"
                },
                "gender": {
                    "type": "boolean"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "moderator_login": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
        "ds.FraxResolveRequest": {
            "type": "object",
            "required": [
                "action"
            ],
            "properties": {
                "action": {
                    "description": "\"complete\" | \"reject\"",
                    "type": "string"
                }
            }
        },
        "ds.FraxUpdateRequest": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "gender": {
                    "type": "boolean"
                },
                "height": {
                    "type": "integer"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}